backend/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
python-dotenv
pymongo
